Added an index to derived_site_name column in historyRecord table.
I made this change live on the table, and it had a dramatic impact
on run time. An full update now runs in 1.5 hours, while it was
taking 2-3 days previously.

Updated the table format in recordsTableWithIndexes and also
updated the create table syntax in initializeRecordsTable.sql.
However, the latter is not tested, since I ran the change live
with SequelPro from my desktop.